Add search customer to  one reseller, and add role admin

+++ src/Events/DeleteCustomerListener.php
@@ -7,16 +7,17 @@ use Symfony\Component\HttpKernel\KernelEvents;
 use Symfony\Component\HttpKernel\Event\ViewEvent;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use ApiPlatform\Core\EventListener\EventPriorities;
+use App\Entity\Customer;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Doctrine\ORM\EntityManagerInterface;
-use App\Entity\Reseller;
-use App\Entity\Customer;
+use Symfony\Component\HttpFoundation\Request;
 
 final class DeleteCustomerListener implements EventSubscriberInterface
 {
 
     private EntityManagerInterface $em;
+
     public function __construct(Security $security, EntityManagerInterface $em)
     {
         $this->security = $security;
@@ -30,36 +31,29 @@ final class DeleteCustomerListener implements EventSubscriberInterface
         ];
     }
 
-    public function checkUserDeleted(ViewEvent $event)
+    public function checkUserDeleted(ViewEvent $event): void
     {
+        $method = $event->getRequest()->getMethod();
         $customer = $event->getControllerResult();
+        $reseller = $this->security->getUser();      
 
-        $resellerRole = $this->security->getUser()->getRoles();
-        $customerRole = $customer->getRoles();
-
-        $reseller = $this->security->getUser();
-
-        if ($customer->getResellers()->getId() !== $reseller->getId())
+        if ($customer instanceof Customer && $customer->getcustomersResellers()->getId() !== $reseller->getId())
+        {throw new AccessDeniedException("Prohibited operation. You can not delete other customer  if is not your.");
+        }elseif($reseller->getRoles() !== ['ROLE_RESELLER']) 
+        {throw new AccessDeniedException("Prohibited operation. You can only delete a customer defined with the property ROLE_USER.");
+        }elseif( Request::METHOD_DELETE === $method)
         {
-            throw new AccessDeniedException("Prohibited operation. You can not delete other customer  if is not your.");
-        }
-        elseif($resellerRole !== ['ROLE_RESELLER'] && $customerRole !== ['ROLE_USER']) 
-        {
-            throw new AccessDeniedException("Prohibited operation. You can only delete a customer defined with the property ROLE_USER.");
-        }
-        elseif($event->getRequest()->getMethod() !== "DELETE")
-        {
-            throw new AccessDeniedException("Prohibited operation. You can only delete a customer with the method DELETE.");
-        }
+            $response = new JsonResponse(
+                ['message' => 'Customer has deleted !'],
+                '200',
+                ['Content-Type' => 'application/json']
+            ); 
 
-        $response = new JsonResponse(
-            ['message' => 'Customer has deleted !'],
-            '200',
-            ['Content-Type' => 'application/json']
-        );       
-        
-        $event->setResponse($response);
-        $this->em->remove($customer);
-        $this->em->flush();
+            $event->setResponse($response);
+            $this->em->remove($customer);
+            $this->em->flush();
+        }      
+        return;
+
     }
 }
